v4: visualization_of_results
* Add Stateful on the app to save the selectioned captor
between 2_Predictions_graph and 3_Box_Plot.
* Delete Whitespaces
* Truncate values in the dataframes
* Move streamlit files from root project to src/
* Delete unused imports
