+++ src/Home.py
@@ -22,12 +22,12 @@ st.markdown(
 )
 
 st.markdown("""
-## Application of Federated Learning to urban mobility data. 
+## Application of Federated Learning to urban mobility data.
 ### Associate to Toulouse city Vilagil Project.
 
 Ongoing Project: Vilagil - Data and mobility.
 
-Applications of Federated Learning to urban mobility data. 
+Applications of Federated Learning to urban mobility data.
 
 
 https://www.banquedesterritoires.fr/sites/default/files/2020-11/Toulouse%2C%20Vilagil%20%28Occitanie%29.pdf
