+++ src/pages/4_Comparison_Models.py
@@ -13,8 +13,7 @@ import numpy as np
 import plotly.express as px
 
 
-from src.metrics import rmse
-from src.config import Params
+from config import Params
 
 
 @st.cache_data
@@ -56,7 +55,7 @@ def compute_absolute_error(path):
 def plot_slider(experiment_path):
     ae_model_1 = compute_absolute_error(experiment_path[0])
     ae_model_2 = compute_absolute_error(experiment_path[1])
-    
+
     def plot_box(title, ae, max_y_value, color):
         fig = px.box(y=ae, color_discrete_sequence=[color], title=title, points="suspectedoutliers")
         fig.update_layout(
@@ -76,13 +75,13 @@ def plot_slider(experiment_path):
         return fig
 
     max_y_value = max(max(ae_model_1), max(ae_model_2))
-    
+
     title_model_1 = paths_experiment_selected[0].split("\\")[1]
     fig_model_1 = plot_box(f"{title_model_1}", ae_model_1, max_y_value, "green")
 
     title_model_2 = paths_experiment_selected[1].split("\\")[1]
     fig_model_2 = plot_box(f"{title_model_2}", ae_model_2, max_y_value, "red")
-    
+
     with st.spinner('Plotting...'):
         st.subheader(f"Comparison between two models on capor {captor} on the federated version (Aboslute Error)")
         _, c2_fed_fig, c3_local_fig, _ = st.columns((1,1,1,1))
@@ -100,11 +99,11 @@ def map_path_experiments_to_params(path_files, params_config_use_for_select):
 
     Parameters:
     -----------
-        path_files : 
+        path_files :
             The path to the config.json of all experiments
         params_config_use_for_select :
             The parameters use for the selection
-            
+
     Returns:
         The mapping between path to the experiments and parameters of the experiements
         exemple :
@@ -134,18 +133,18 @@ def map_path_experiments_to_params(path_files, params_config_use_for_select):
 
 def selection_of_experiment(possible_choice):
     time_serie_percentage_length = st.selectbox('Choose the time series length', possible_choice["time_serie_percentage_length"].keys())
-    
+
     nb_captor_filtered = filtering_path_file(possible_choice["number_of_nodes"], possible_choice["time_serie_percentage_length"][time_serie_percentage_length])
     nb_captor = st.selectbox('Choose the number of captor', nb_captor_filtered.keys())
 
     windows_size_filtered = filtering_path_file(possible_choice["window_size"], possible_choice["number_of_nodes"][nb_captor])
     window_size = st.selectbox('Choose the windows size', windows_size_filtered.keys())
-        
+
     horizon_filtered = filtering_path_file(possible_choice["prediction_horizon"], windows_size_filtered[window_size])
     horizon_size = st.selectbox('Choose the prediction horizon', horizon_filtered.keys())
-    
+
     models_filtered = filtering_path_file(possible_choice["model"], horizon_filtered[horizon_size])
-    
+
     col1_model_1, col2_model_2 = st.columns(2)
     with col1_model_1:
         model_1 = st.radio(
@@ -168,7 +167,7 @@ st.header("Comparison Models")
 
 experiments = "experiments" # PATH where your experiments are saved
 if path_files := glob.glob(f"./{experiments}/**/config.json", recursive=True):
-    
+
     params_config_use_for_select = \
     [
         "time_serie_percentage_length",
@@ -220,19 +219,19 @@ if path_files := glob.glob(f"./{experiments}/**/config.json", recursive=True):
         federated_node_model_2 = pd.DataFrame(federated_node_model_2, columns=multiselect_metrics, index=["Captor in Federation"])
 
     _, c2_title_df, _ = st.columns((2,1,2))
-    
+
     with c2_title_df:
         st.header("Captor in Federation")
-    
+
     c1_model_1, c2_model_2 = st.columns(2)
     with c1_model_1:
         model_1_name = paths_experiment_selected[0].split("\\")[1]
         st.subheader(f"{model_1_name}")
-        st.table(federated_node_model_1.style.set_table_styles([{'selector': 'th', 'props': [('font-weight', 'bold'), ('color', 'black')]}]))
+        st.table(federated_node_model_1.style.set_table_styles([{'selector': 'th', 'props': [('font-weight', 'bold'), ('color', 'black')]}]).format("{:.2f}"))
     with c2_model_2:
         model_2_name = paths_experiment_selected[1].split("\\")[1]
         st.subheader(f"{model_2_name}")
-        st.table(federated_node_model_2.style.set_table_styles([{'selector': 'th', 'props': [('font-weight', 'bold'), ('color', 'black')]}]))
+        st.table(federated_node_model_2.style.set_table_styles([{'selector': 'th', 'props': [('font-weight', 'bold'), ('color', 'black')]}]).format("{:.2f}"))
 
 
     params_model_1 = Params(f'{path_model_1}/config.json')
